Ryddet i duplisert code iflg sonar
