+++ src/felles-komponenter/skjema/datovelger/utils.ts
@@ -4,7 +4,7 @@ import { erGyldigISODato, toDatePrettyPrint } from '../../../utils';
 
 export const validerDato = (value: string | null, tidligsteFom?: string, senesteTom?: string) => {
     if (!value || value.trim().length === 0) {
-        return undefined;
+        return;
     }
 
     if (!erGyldigISODato(value)) {
@@ -25,6 +25,4 @@ export const validerDato = (value: string | null, tidligsteFom?: string, seneste
 
         return `Datoen må være innenfor perioden ${prettyFra}-${prettyTil}`;
     }
-
-    return undefined;
 };
