+++ src/moduler/aktivitet/aktivitet-forms/behandling-gammel/validate.js
@@ -2,7 +2,7 @@ import { validerDato } from '../../../../felles-komponenter/skjema/datovelger/ut
 
 export function validateBehandlingType(avtalt, value) {
     if (avtalt) {
-        return undefined;
+        return;
     }
     if (!value || value.trim().length === 0) {
         return 'Du må fylle ut type behandling';
@@ -10,8 +10,6 @@ export function validateBehandlingType(avtalt, value) {
     if (value.length > 100) {
         return `Du må korte ned teksten til 100 tegn`;
     }
-
-    return null;
 }
 export function validateBehandlingSted(avtalt, value) {
     if (avtalt) {
@@ -23,18 +21,15 @@ export function validateBehandlingSted(avtalt, value) {
     if (value.length > 255) {
         return `Du må korte ned teksten til 255 tegn`;
     }
-
-    return undefined;
 }
 
 export function validateFeltForLangt(avtalt, value) {
     if (avtalt) {
-        return undefined;
+        return;
     }
     if (value && value.length > 255) {
         return `Du må korte ned teksten til 255 tegn`;
     }
-    return undefined;
 }
 
 export function validateFraDato(avtalt, tilDato, value) {
