+++ src/moduler/aktivitet/aktivitet-forms/egen/validate.js
@@ -2,7 +2,7 @@ import { validerDato } from '../../../../felles-komponenter/skjema/datovelger/ut
 
 export function validateTittel(avtalt, value) {
     if (avtalt) {
-        return undefined;
+        return;
     }
     if (!value || value.trim().length <= 0) {
         return 'Du må fylle ut navn på aktiviteten';
@@ -10,48 +10,6 @@ export function validateTittel(avtalt, value) {
     if (value.length > 100) {
         return `Du må korte ned teksten til 100 tegn`;
     }
-
-    return undefined;
-}
-
-export function validateFeltForLangt(avtalt, value) {
-    if (avtalt) {
-        return undefined;
-    }
-    if (value && value.length > 255) {
-        return `Du må korte ned teksten til 255 tegn`;
-    }
-    return undefined;
-}
-
-export function validateBeskrivelse(avtalt, value) {
-    if (avtalt) {
-        return undefined;
-    }
-    if (value && value.length > 5000) {
-        return `Du må korte ned teksten til 5000 tegn`;
-    }
-    return undefined;
-}
-
-export function validateLenke(avtalt, value) {
-    if (avtalt) {
-        return undefined;
-    }
-    if (value && value.length > 2000) {
-        return `Du må korte ned teksten til 2000 tegn`;
-    }
-    return undefined;
-}
-
-export function validateFraDato(avtalt, tilDato, value) {
-    if (avtalt) {
-        return undefined;
-    }
-    if (!value || value.trim().length <= 0) {
-        return 'Du må fylle ut fra dato';
-    }
-    return validerDato(value, tilDato, null);
 }
 
 export function validateTilDato(fraDato, value) {
