+++ src/moduler/aktivitet/aktivitet-forms/sokeavtale/validate.js
@@ -2,38 +2,35 @@ import { validerDato } from '../../../../felles-komponenter/skjema/datovelger/ut
 
 export function validateBeskrivelse(avtalt, value) {
     if (avtalt) {
-        return undefined;
+        return;
     }
     if (value && value.length > 5000) {
         return `Du må korte ned teksten til 5000 tegn`;
     }
-    return undefined;
 }
 
 export function validateOppfolging(avtalt, value) {
     if (avtalt) {
-        return undefined;
+        return;
     }
     if (value && value.length > 255) {
         return `Du må korte ned teksten til 255 tegn`;
     }
-    return undefined;
 }
 
 export function validateAntallStillinger(avtalt, value) {
     if (avtalt || !value) {
-        return undefined;
+        return;
     }
 
     if (!Number.isInteger(Number(value))) {
         return 'Antall må være et heltall';
     }
-    return undefined;
 }
 
 export function validateAntallStillingerIUken(avtalt, value, antallStillingerSokes) {
     if (avtalt || !!antallStillingerSokes) {
-        return undefined;
+        return;
     }
 
     if (value.length === 0) {
@@ -43,7 +40,6 @@ export function validateAntallStillingerIUken(avtalt, value, antallStillingerSok
     if (!Number.isInteger(Number(value))) {
         return 'Antall må være et heltall';
     }
-    return undefined;
 }
 
 export function validateFraDato(avtalt, tilDato, value) {
